[WebAssembly] Emit `__llvm_covmap` and `__llvm_covfun` as custom sections
This patch makes `__llvm_covmap` and `__llvm_covfun` as custom sections
in the wasm object file because they will not be referenced at runtime
but just used by `llvm-cov` post-processing tools. The same approach is
used in the ELF object file to emit them without `SHF_ALLOC` flag.

Those sections have their associated comdat group symbols, which are
section symbols (not data symbols) in the wasm object file. This patch
also sets the symbol type explicitly not to hit an assertion in the
`WasmObjectWriter::writeOneObject` ("data symbols must live in a data
section").