[wasm-ld][Coverage] Align `__llvm_{covfun,covmap}` sections to 8 bytes
This patch aligns the `__llvm_{covfun,covmap}` sections to 8 bytes at
the linking stage. This is required because llvm-cov assumes that
records in the sections are 8-byte aligned. Unfortunately, custom
sections does not have an alignment attribute in the WebAssembly linking
spec, so we temporarily do a special case for the metadata sections.
In the future, we should generalize this to support alignment for
relocatable custom sections.